requirements.txt
typer[all]
python-frontmatter
markdown-it-py
pyyaml
python-dotenv
anthropic
chromadb
openai
rich
